- Implement My Question show full image

+++ Otta/View/OttaQuestionFeedView/OttaQuestionFeedCell.h
@@ -46,7 +46,7 @@
 - (IBAction)imageBtnTapped:(id)sender;
 - (void) startPerformSelectorForDelete;
 
-- (void) selectAnswerIndex:(int)answerIndex;
+- (void) selectAnswerIndex:(NSInteger)answerIndex;
 - (void) deselectCell;
 
 @end
